Release v2.6.5

+++ framework/core/src/include/kungfu/yijinjing/io.h
@@ -53,15 +53,15 @@ public:
 
   [[maybe_unused]] journal::reader_ptr open_reader(const data::location_ptr &location, uint32_t dest_id);
 
-  journal::writer_ptr open_writer(uint32_t dest_id, uint32_t page_size = 0);
+  journal::writer_ptr open_writer(uint32_t dest_id, uint64_t page_size = 0);
 
-  journal::writer_ptr open_writer_at(const data::location_ptr &location, uint32_t dest_id, uint32_t page_size = 0);
+  journal::writer_ptr open_writer_at(const data::location_ptr &location, uint32_t dest_id, uint64_t page_size = 0);
 
   journal::writer_ptr open_hookable_writer(uint32_t dest_id, const journal::writer_hook_ptr &hook,
-                                           uint32_t page_size = 0);
+                                           uint64_t page_size = 0);
 
   journal::writer_ptr open_hookable_writer_at(const data::location_ptr &location, uint32_t dest_id,
-                                              const journal::writer_hook_ptr &hook, uint32_t page_size = 0);
+                                              const journal::writer_hook_ptr &hook, uint64_t page_size = 0);
 
   [[nodiscard]] nanomsg::url_factory_ptr get_url_factory() const { return url_factory_; }
