+++ framework/core/src/include/kungfu/yijinjing/journal/page.h
@@ -18,7 +18,7 @@ public:
 
   [[nodiscard]] uint32_t get_version() const { return header_->version; }
 
-  [[nodiscard]] uint32_t get_page_size() const { return header_->page_size; }
+  [[nodiscard]] uint64_t get_page_size() const { return header_->page_size; }
 
   [[nodiscard]] data::location_ptr get_location() const { return location_; }
 
@@ -40,7 +40,7 @@ public:
     return address() + header_->page_size - sizeof(longfist::types::frame_header);
   }
 
-  [[nodiscard]] uint32_t get_body_size() const { return size_ - header_->page_header_length; }
+  [[nodiscard]] uint64_t get_body_size() const { return size_ - header_->page_header_length; }
 
   [[nodiscard]] uintptr_t first_frame_address() const { return address() + header_->page_header_length; }
 
@@ -51,7 +51,7 @@ public:
            address_border();
   }
 
-  static page_ptr load(const data::location_ptr &location, uint32_t dest_id, uint32_t page_size, uint32_t page_id,
+  static page_ptr load(const data::location_ptr &location, uint32_t dest_id, uint64_t page_size, uint32_t page_id,
                        bool is_writing, bool lazy, bool pre_open = false);
 
   static page_ptr load_header_and_1st_frame_header(const data::location_ptr &location, uint32_t dest_id,
@@ -61,7 +61,7 @@ public:
 
   static uint32_t find_page_id(const data::location_ptr &location, uint32_t dest_id, int64_t time);
 
-  static uint32_t find_page_size(const data::location_ptr &location, uint32_t dest_id, uint32_t page_size = 0);
+  static uint64_t find_page_size(const data::location_ptr &location, uint32_t dest_id, uint64_t page_size = 0);
 
   static bool check_page_existed(const data::location_ptr &location, uint32_t dest_id);
