+++ framework/core/src/libkungfu/yijinjing/practice/hero.cpp
@@ -373,7 +373,7 @@ void hero::require_write_to(int64_t trigger_time, uint32_t source_id, uint32_t d
 }
 
 void hero::require_write_to_band(int64_t trigger_time, uint32_t source_id, const data::location_ptr &location,
-                                 uint32_t page_size) const {
+                                 uint64_t page_size) const {
   auto writer = get_writer(source_id);
   RequestWriteToBand msg = {};
   location->to<RequestWriteToBand>(msg);
